[PI-230] Renamed postgres-consistency package to importer-db-consistency
