+++ importer-db-consistency/README.md
@@ -1,4 +1,4 @@
-# `cardano-sl-postgres-consistency`
+# `cardano-sl-importer-db-consistency`
 
 Checks the consistency of the postgres db generated by the `cardano-sl-blockchain-importer`
