+++ importer-db-consistency/scripts/build.sh
