+++ importer-db-consistency/scripts/fetch_dependencies.sh
