+++ importer-db-consistency/scripts/launch/external-consistency-from-blk.sh
@@ -5,37 +5,31 @@
 #   - Having the postgres db up-to-date
 #   - Configuring the environment values for the postgres db (i.e.: source setup-localDB.sh)
 # Usage:
-#   ./random-external-consistency.sh CHAIN TOP_EPOCH NUMBER_BLOCKS KV_DB_LOCATION
+#   ./external-consistency-from-blk.sh CHAIN KV_DB_LOCATION STARTING_BLK
 
 scriptDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 repoDir="${scriptDir}/../../.."
-blkFile="${repoDir}/postgres-consistency/blkHashes.txt"
-logsFile="${repoDir}/postgres-consistency/externalConsistency.log"
+logsFile="${repoDir}/importer-db-consistency/externalConsistency.log"
 topologyFile="/tmp/external-topology.yaml"
 
 . ${scriptDir}/utils.sh
 
 # Parameters
 chain="$1"
-topEpoch="$2"
-numberBlocks="$3"
-kvDBLocation="$4"
+kvDBLocation="$2"
+startingBlk="$3"
 
 CONFIG_KEY=
 KEY_FILE=
 TOPOLOGY_HOST=
 setup_chain_config ${chain}
 
-# FIXME: Do npm install node-fetch?
 logWithTimestamp "Doing setup"
-${repoDir}/scripts/build/cardano-sl.sh postgres-consistency > /dev/null
+${repoDir}/scripts/build/cardano-sl.sh importer-db-consistency > /dev/null
 printf "wallet:\n relays: [[{ host: ${TOPOLOGY_HOST} }]]\n valency: 1\n fallbacks: 7" > ${topologyFile}
 
-logWithTimestamp "Getting random blocks to check"
-node ${scriptDir}/../EpochSlotToBlkHash.js ${topEpoch} ${numberBlocks} > ${blkFile}
-
-logWithTimestamp "Running external consistency test"
-stack exec -- cardano-postgres-consistency ext-const-random --blocks-file ${blkFile} \
+logWithTimestamp "Running external consistency test from blk"
+stack exec -- cardano-importer-db-consistency ext-const-from-blk --starting-block ${startingBlk} \
            --topology "${topologyFile}" \
            --log-config "${repoDir}/blockchain-importer/log-config.yaml" \
            --logs-prefix "${repoDir}/external-logs" \
@@ -47,6 +41,3 @@ stack exec -- cardano-postgres-consistency ext-const-random --blocks-file ${blkF
            --postgres-host ${DB_HOST} --postgres-port ${DB_PORT} > ${logsFile}
 
 check_succeded_on_logs "External consistency" ${logsFile}
-
-# Clean up
-rm ${blkFile}
