+++ importer-db-consistency/scripts/launch/random-external-consistency.sh
@@ -5,31 +5,37 @@
 #   - Having the postgres db up-to-date
 #   - Configuring the environment values for the postgres db (i.e.: source setup-localDB.sh)
 # Usage:
-#   ./external-consistency-from-blk.sh CHAIN KV_DB_LOCATION STARTING_BLK
+#   ./random-external-consistency.sh CHAIN TOP_EPOCH NUMBER_BLOCKS KV_DB_LOCATION
 
 scriptDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 repoDir="${scriptDir}/../../.."
-logsFile="${repoDir}/postgres-consistency/externalConsistency.log"
+blkFile="${repoDir}/importer-db-consistency/blkHashes.txt"
+logsFile="${repoDir}/importer-db-consistency/externalConsistency.log"
 topologyFile="/tmp/external-topology.yaml"
 
 . ${scriptDir}/utils.sh
 
 # Parameters
 chain="$1"
-kvDBLocation="$2"
-startingBlk="$3"
+topEpoch="$2"
+numberBlocks="$3"
+kvDBLocation="$4"
 
 CONFIG_KEY=
 KEY_FILE=
 TOPOLOGY_HOST=
 setup_chain_config ${chain}
 
+# FIXME: Do npm install node-fetch?
 logWithTimestamp "Doing setup"
-${repoDir}/scripts/build/cardano-sl.sh postgres-consistency > /dev/null
+${repoDir}/scripts/build/cardano-sl.sh importer-db-consistency > /dev/null
 printf "wallet:\n relays: [[{ host: ${TOPOLOGY_HOST} }]]\n valency: 1\n fallbacks: 7" > ${topologyFile}
 
-logWithTimestamp "Running external consistency test from blk"
-stack exec -- cardano-postgres-consistency ext-const-from-blk --starting-block ${startingBlk} \
+logWithTimestamp "Getting random blocks to check"
+node ${scriptDir}/../EpochSlotToBlkHash.js ${topEpoch} ${numberBlocks} > ${blkFile}
+
+logWithTimestamp "Running external consistency test"
+stack exec -- cardano-importer-db-consistency ext-const-random --blocks-file ${blkFile} \
            --topology "${topologyFile}" \
            --log-config "${repoDir}/blockchain-importer/log-config.yaml" \
            --logs-prefix "${repoDir}/external-logs" \
@@ -41,3 +47,6 @@ stack exec -- cardano-postgres-consistency ext-const-from-blk --starting-block $
            --postgres-host ${DB_HOST} --postgres-port ${DB_PORT} > ${logsFile}
 
 check_succeded_on_logs "External consistency" ${logsFile}
+
+# Clean up
+rm ${blkFile}
