+++ importer-db-consistency/scripts/launch/setup-localDB.sh
@@ -1,5 +1,5 @@
 export DB_USER="postgres"
 export DB_HOST="localhost"
-export DB="icaruspocbackendservice"
+export DB="mainnet"
 export DB_PASSWORD="mysecretpassword"
 export DB_PORT=5432
