+++ importer-db-consistency/scripts/launch/utils.sh
