+++ importer-db-consistency/src/importer-db-consistency/ImporterDBConsistencyNodeOptions.hs
@@ -4,11 +4,11 @@
 
 -- | Command line options of blockchainImporter node.
 
-module PostgresConsistencyNodeOptions
-       ( PostgresConsistencyNodeArgs (..)
-       , PostgresConsistencyArgs (..)
+module ImporterDBConsistencyNodeOptions
+       ( ImporterDBConsistencyNodeArgs (..)
+       , ImporterDBConsistencyArgs (..)
        , PostgresChecks (..)
-       , getPostgresConsistencyNodeOptions
+       , getImporterDBConsistencyNodeOptions
        ) where
 
 import           Universum
@@ -24,9 +24,9 @@ import           Pos.Client.CLI (CommonNodeArgs (..))
 import qualified Pos.Client.CLI as CLI
 
 
-data PostgresConsistencyNodeArgs = PostgresConsistencyNodeArgs
-    { enaCommonNodeArgs          :: !CommonNodeArgs
-    , enaPostgresConsistencyArgs :: !PostgresConsistencyArgs
+data ImporterDBConsistencyNodeArgs = ImporterDBConsistencyNodeArgs
+    { enaCommonNodeArgs            :: !CommonNodeArgs
+    , enaImporterDBConsistencyArgs :: !ImporterDBConsistencyArgs
     } deriving Show
 
 data PostgresChecks = ExternalConsistencyFromBlk String
@@ -36,8 +36,8 @@ data PostgresChecks = ExternalConsistencyFromBlk String
                     | GetTipHash
                     deriving Show
 
--- | PostgresConsistency specific arguments.
-data PostgresConsistencyArgs = PostgresConsistencyArgs
+-- | ImporterDBConsistency specific arguments.
+data ImporterDBConsistencyArgs = ImporterDBConsistencyArgs
     { webPort             :: !Word16
     -- ^ The port for the blockchainImporter backend
     , postGresConfig      :: !PGS.ConnectInfo
@@ -78,7 +78,7 @@ connectInfoParser = do
   pure PGS.ConnectInfo{..}
 
 -- | Ther parser for the blockchainImporter arguments.
-blockchainImporterArgsParser :: Parser PostgresConsistencyNodeArgs
+blockchainImporterArgsParser :: Parser ImporterDBConsistencyNodeArgs
 blockchainImporterArgsParser = do
     commonNodeArgs <- CLI.commonNodeArgsParser
     webPort        <- CLI.webPortOption 8200 "Port for web API."
@@ -89,7 +89,7 @@ blockchainImporterArgsParser = do
         value   0 <>
         help    "First block whose info will be stored on postgres DB."
     checksToDo <- postgresCheckParser
-    pure $ PostgresConsistencyNodeArgs commonNodeArgs PostgresConsistencyArgs{..}
+    pure $ ImporterDBConsistencyNodeArgs commonNodeArgs ImporterDBConsistencyArgs{..}
 
 postgresCheckParser :: Parser PostgresChecks
 postgresCheckParser = do
@@ -134,8 +134,8 @@ postgresCheckParser = do
 
 
 -- | The parser for the blockchainImporter.
-getPostgresConsistencyNodeOptions :: IO PostgresConsistencyNodeArgs
-getPostgresConsistencyNodeOptions = execParser programInfo
+getImporterDBConsistencyNodeOptions :: IO ImporterDBConsistencyNodeArgs
+getImporterDBConsistencyNodeOptions = execParser programInfo
   where
     programInfo = info (helper <*> versionOption <*> blockchainImporterArgsParser) $
         fullDesc <> progDesc "Consistency checker of the postgres db generated by blockchain importer."
@@ -143,5 +143,5 @@ getPostgresConsistencyNodeOptions = execParser programInfo
                  <> footerDoc CLI.usageExample
 
     versionOption = infoOption
-        ("cardano-postgres-consistency-" <> showVersion version)
+        ("cardano-importer-db-consistency-" <> showVersion version)
         (long "version" <> help "Show version.")
